sh 수정

+++ myprog.sh
@@ -1,10 +1,31 @@
 #! /bin/bash
-mkdir templates
-mv cos_result.html templates
-mv popup.html templates
-mv result.html templates
-mv upload.html templates
-mv tf_result.html templates
+
+if [ ! -d "templates" ]; then
+	mkdir templates
+fi
+
+if [ -f "cos_result.html" ]; then
+	mv cos_result.html templates
+fi
+
+if [ -f "popup.html" ]; then
+	mv popup.html templates
+fi
+
+if [ -f "result.html" ]; then
+	mv result.html templates
+fi
+
+if [ -f "upload.html" ]; then
+	mv upload.html templates
+fi
+
+if [ -f "tf_result.html" ]; then
+	mv tf_result.html templates
+fi
+
+chmod 777 final.py
+
 ./final.py
